[REFACTORING] added packages to group unit tests
